added step def java file and cleaned up sample step def
